@@ -17,13 +17,13 @@
     }
   ],
   "require": {
-    "php": "^7.1|^8.0",
-    "symfony/translation-contracts": "^1.1|^2.3|^3.0",
-    "symfony/form": "^4.4|^5.2|^6.0",
-    "symfony/framework-bundle": "^4.4|^5.2|^6.0",
-    "symfony/validator": "^4.4|^5.2|^6.0",
-    "doctrine/doctrine-bundle": "^1.6|^2.1",
-    "doctrine/orm": "^2.5"
+    "php": "^8.1",
+    "symfony/translation-contracts": "^3.0",
+    "symfony/form": "^5.4|^6.0",
+    "symfony/framework-bundle": "^5.4|^6.0",
+    "symfony/validator": "^5.4|^6.0",
+    "doctrine/doctrine-bundle": "^2.1",
+    "doctrine/orm": "^3.0"
   },
   "suggest": {
     "umanit/translation-bundle": "Allow your panel to be translated."
