@@ -13,12 +13,8 @@ use Umanit\BlockBundle\Model\PanelInterface;
 
 class PanelDataTransformer implements DataTransformerInterface
 {
-    /** @var EntityManagerInterface */
-    private $em;
-
-    public function __construct(EntityManagerInterface $em)
+    public function __construct(private readonly EntityManagerInterface $em)
     {
-        $this->em = $em;
     }
 
     public function transform($value): ?array
