@@ -9,35 +9,17 @@ use Umanit\BlockBundle\Entity\Panel;
 
 trait BlockTrait
 {
-    /**
-     * @var int
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
     #[ORM\Column(name: 'id', type: 'integer')]
     #[ORM\Id]
     #[ORM\GeneratedValue(strategy: 'AUTO')]
-    private $id;
+    private ?int $id;
 
-    /**
-     * @var PanelInterface
-     *
-     * @ORM\ManyToOne(targetEntity="Umanit\BlockBundle\Entity\Panel")
-     * @ORM\JoinColumn(name="panel_id", referencedColumnName="id", onDelete="CASCADE")
-     */
     #[ORM\ManyToOne(targetEntity: Panel::class)]
     #[ORM\JoinColumn(name: 'panel_id', referencedColumnName: 'id', onDelete: 'CASCADE')]
-    protected $panel;
+    protected ?PanelInterface $panel;
 
-    /**
-     * @var int
-     *
-     * @ORM\Column(name="position", type="integer")
-     */
     #[ORM\Column(name: 'position', type: 'integer')]
-    protected $position;
+    protected ?int $position;
 
     public function getPanel(): PanelInterface
     {
