@@ -12,23 +12,11 @@ use Umanit\BlockBundle\Resolver\BlockManagerResolver;
 
 class BlockRuntime implements RuntimeExtensionInterface
 {
-    /** @var BlockManagerResolver */
-    private $blockManagerResolver;
-
-    /** @var bool */
-    private $debugIsEnabled;
-
-    /** @var LoggerInterface */
-    private $logger;
-
     public function __construct(
-        BlockManagerResolver $blockManagerResolver,
-        LoggerInterface $logger,
-        $debugIsEnabled = false
+        private readonly BlockManagerResolver $blockManagerResolver,
+        private readonly LoggerInterface $logger,
+        private readonly bool $debugIsEnabled = false
     ) {
-        $this->blockManagerResolver = $blockManagerResolver;
-        $this->logger = $logger;
-        $this->debugIsEnabled = $debugIsEnabled;
     }
 
     /**
@@ -39,6 +27,7 @@ class BlockRuntime implements RuntimeExtensionInterface
      *
      * @return string
      * @throws BlockManagerNotFoundException
+     * @throws \Exception
      */
     public function renderBlock(BlockInterface $block, array $parameters = []): string
     {
